+++ source/config/configSpec.py
@@ -228,8 +228,8 @@ schemaVersion = integer(min=0, default={latestSchemaVersion})
 	reportLineSpacing = boolean(default=false)
 	reportStyle = boolean(default=false)
 	# Bitwise combination of none, some or all values of ReportSpellingErrors
-	# 1: Speech, 2: Sound
-	reportSpellingErrors2 = integer(min=0, max=3, default=1)
+	# 1: Speech, 2: Sound, 4: Braille
+	reportSpellingErrors2 = integer(min=0, max=7, default=1)
 	reportPage = boolean(default=true)
 	reportLineNumber = boolean(default=False)
 	# 0: Off, 1: Speech, 2: Tones, 3: Both Speech and Tones